+++ example/lib/main.dart
@@ -91,7 +91,7 @@ class _PlayoutExampleState extends State<PlayoutExample> {
               "AV Player",
               style: Theme.of(context)
                   .textTheme
-                  .title
+                  .headline6
                   .copyWith(color: Colors.white),
             )
           ],
@@ -106,7 +106,7 @@ class _PlayoutExampleState extends State<PlayoutExample> {
                 padding: EdgeInsets.fromLTRB(17.0, 33.0, 17.0, 0.0),
                 child: Text(
                   "Video Player",
-                  style: Theme.of(context).textTheme.display1.copyWith(
+                  style: Theme.of(context).textTheme.headline4.copyWith(
                       color: Colors.pink[500], fontWeight: FontWeight.w600),
                 ),
               ),
@@ -116,7 +116,7 @@ class _PlayoutExampleState extends State<PlayoutExample> {
                 padding: EdgeInsets.fromLTRB(17.0, 0.0, 17.0, 30.0),
                 child: Text(
                   "Plays video from a URL with background audio support and lock screen controls.",
-                  style: Theme.of(context).textTheme.subhead.copyWith(
+                  style: Theme.of(context).textTheme.subtitle1.copyWith(
                       color: Colors.white70, fontWeight: FontWeight.w400),
                 ),
               ),
@@ -131,7 +131,7 @@ class _PlayoutExampleState extends State<PlayoutExample> {
                 padding: EdgeInsets.fromLTRB(17.0, 23.0, 17.0, 0.0),
                 child: Text(
                   "Audio Player",
-                  style: Theme.of(context).textTheme.display1.copyWith(
+                  style: Theme.of(context).textTheme.headline4.copyWith(
                       color: Colors.pink[500], fontWeight: FontWeight.w600),
                 ),
               ),
@@ -141,7 +141,7 @@ class _PlayoutExampleState extends State<PlayoutExample> {
                 padding: EdgeInsets.fromLTRB(17.0, 0.0, 17.0, 30.0),
                 child: Text(
                   "Plays audio from a URL with background audio support and lock screen controls.",
-                  style: Theme.of(context).textTheme.subhead.copyWith(
+                  style: Theme.of(context).textTheme.subtitle1.copyWith(
                       color: Colors.white70, fontWeight: FontWeight.w400),
                 ),
               ),